[Feature Enhancement] Add resumable processing support to OpNamesExtractor
