+++ graph_net/tools/typical_sequence_decompose.sh
@@ -10,7 +10,7 @@ model_list="$GRAPH_NET_ROOT/graph_net/config/small10_torch_samples_list.txt"
 
 python3 -m graph_net.apply_sample_pass \
     --model-path-list $model_list \
-    --sample-pass-file-path $GRAPH_NET_ROOT/graph_net/torch/op_names_extractor.py \
+    --sample-pass-file-path $GRAPH_NET_ROOT/graph_net/torch/sample_pass/op_names_extractor.py \
     --sample-pass-class-name OpNamesExtractor \
     --sample-pass-config=$(base64 -w 0 <<EOF
 {
